+++ frontend/package.json
@@ -3,10 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "bootstrap": "^4.3.1",
     "normalize.css": "^8.0.1",
     "react": "^16.11.0",
+    "react-bootstrap": "^1.0.0-beta.14",
     "react-dom": "^16.11.0",
-    "react-modal": "^3.11.1"
+    "react-modal": "^3.11.1",
+    "react-router-dom": "^5.1.2"
   },
   "devDependencies": {
     "react-scripts": "3.2.0"
