+++ frontend/src/components/Header.jsx
@@ -3,14 +3,10 @@ import React, { Component } from 'react';
 class Header extends Component {
   constructor(props) {
     super(props);
-    this.state = {
-    };
+    this.state = {};
   }
 
-  componentDidMount() {
-
-  }
-
+  componentDidMount() {}
 
   render() {
     return (
