+++ frontend/src/components/Profile.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+class Profile extends Component {
+  render() {
+    return (
+      <div>
+        <button variant="secondary">Close</button>
+        <button variant="primary">Save changes</button>
+      </div>
+    );
+  }
+}
+
+export default Profile;
